\name{SignifReg-package}
\alias{SignifReg-package}
%\alias{SignifReg}
\docType{package}
\title{
\packageTitle{SignifReg}
}
\description{
\packageDescription{SignifReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SignifReg}
%\packageIndices{SignifReg}

}
\author{
\packageAuthor{SignifReg}

Maintainer: \packageMaintainer{SignifReg}
}
\references{
Zambom A Z, Kim J. Consistent significance controlled variable selection in high-dimensional regression. Stat.2018;7:e210. https://doi.org/10.1002/sta4.210
}


